refactor: update file names, position and its imports
