+++ rocket-mail/src/functions/GetEmailContents.ts
